api_python/requirements.txt
flask
flask-cors
python-dotenv
transformers
torch
huggingface-hub
groq
pydantic
accelerate
google-genai